FIX: minor bug fixes and extraction.
