+++ src/index.css
@@ -12,11 +12,13 @@
 	--colour-secondary-second: #698dc7;
 	--box-shadow-level-1: 0px 0px 5px 0px rgba(0, 0, 0, 0.2);
 	--box-shadow-level-2: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
+	--colour-text-grey: #2d333a;
 	--box-shadow-level-3: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
 }
 
 [data-theme='dark'] {
 	--colour-secondary-second: #141f30;
+	--colour-text-grey: #8b949e;
 }
 
 .app {
